requirements.txt
paramiko
pycrypto
psutil
MySQL-python